Fix save_and_exit edge case
When the user selects "Save" in the "Lose changes?" menu, if the
filename is not initialized, the user will be prompted for a file name.
If the user then presses ESC to exit this prompt, then we return to the
"Lose changes?" menu.  (Prior to this change it would exit without
saving changes.)